extract methods for board model
